Refactor Sarif issue removers
