+++ src/SonarScanner.MSBuild/IgnoredIssues/SarifIssueRemover10.cs
@@ -1,0 +1,64 @@
+﻿/*
+ * SonarScanner for .NET
+ * Copyright (C) 2016-2022 SonarSource SA
+ * mailto: info AT sonarsource DOT com
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 3 of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+ */
+
+using System.Collections.Generic;
+using System.Linq;
+using Newtonsoft.Json.Linq;
+
+namespace SonarScanner.MSBuild.IgnoredIssues
+{
+    public static class SarifIssueRemover10
+    {
+        private const string Runs = "runs";
+        private const string Results = "results";
+        private const string RuleId = "ruleId";
+        private const string Rules = "rules";
+        private const string Id = "id";
+
+        public static bool Filter(JObject issuesJson, ISet<string> sonarRuleIds)
+        {
+            var tokensToRemove = new List<JToken>();
+            if (!issuesJson.ContainsKey(Runs)
+                || issuesJson.GetValue(Runs).First == null)
+            {
+                return false;
+            }
+
+            var run = issuesJson.GetValue(Runs).First;
+            if (run[Results] != null)
+            {
+                tokensToRemove.AddRange(
+                    run[Results].Where(x => x[RuleId].Value<string>() != null
+                                            && !sonarRuleIds.Contains(x[RuleId].Value<string>())));
+            }
+
+            if (run[Rules] != null)
+            {
+                tokensToRemove.AddRange(
+                    run[Rules].Where(x => x.First != null
+                                          && x.First[Id].Value<string>() != null
+                                          && !sonarRuleIds.Contains(x.First[Id].Value<string>())));
+            }
+
+            tokensToRemove.ForEach(x => x.Remove());
+            return tokensToRemove.Any();
+        }
+    }
+}
